Added unit test and fixed .NET culture stuff

+++ src/Secs4Net/MessageType.cs
@@ -1,6 +1,6 @@
 ﻿namespace Secs4Net;
 
-public enum MessageType : byte
+public enum MessageType : ushort
 {
     DataMessage = 0b0000_0000,
     SelectRequest = 0b0000_0001,
